requirements.txt
# Core Dependencies
pip>=21.0.0
setuptools>=65.5.1
wheel>=0.38.0
nest_asyncio==1.5.8

# API and Exchange
python-binance==1.0.16
ccxt==2.9.11
websockets==10.4
requests==2.31.0
aiohttp==3.8.5

# Web Framework (Fixed versions for compatibility)
fastapi==0.78.0
uvicorn[standard]==0.18.3
pydantic==1.9.2
starlette==0.19.1
python-multipart==0.0.6
jinja2==3.1.2

# Database
pymongo==4.3.3

# Data Processing
numpy==1.21.6
pandas==1.4.4
python-dateutil==2.8.2
pytz==2023.3

# System Monitoring
psutil==5.9.5

# Visualization
plotly==5.13.1

# Configuration
python-dotenv==0.21.1
pyyaml==6.0.1

# Development
pytest==7.3.1
black==23.3.0
flake8==6.0.0